refactor: extract files from main
