+++ docker-gpr/main.js
@@ -1,90 +1,25 @@
 const { exec } = require("@actions/exec");
-const { getInput, setFailed, setOutput, debug, setSecret } = require("@actions/core");
-
-const BUILD_ARGS = ['NPM_TOKEN'];
-
-const parseBuildArgs = () => {
-  const envValueExists = BUILD_ARGS.some(env => !!process.env[env]);
-  if (!envValueExists) { return []; }
-
-  return BUILD_ARGS.filter(env => !!process.env[env])
-    .map(env => (`--build-arg ${env}=${process.env[env]}`));
-}
-
-const getBranchName = () => {
-  const ref = process.env.GITHUB_REF;
-  const pr_ref = process.env.GITHUB_HEAD_REF;
-  debug(ref);
-  debug(pr_ref);
-  if (ref && /\/development|\/master/.test(ref)) {
-    return ref.replace('refs/heads/', '');
-  } else if (pr_ref && /\/pull\//.test(ref)) {
-    // Replace characters that isn't allowed in a docker tag name
-    return pr_ref.replace(/(^\.)|(^-)|[^A-Za-z0-9-_.]/g, '-').slice(0, 128);
-  }
-}
-
-const parseHeadTag = (packagePath) => {
-  const headTag = [];
-  const tagHead = getInput("head-tag").toLowerCase().trim();
-
-  const stageBranch = getInput("stage-branch").toLowerCase().trim();
-  const stageTag = getInput("stage-tag").toLowerCase().trim();
-
-  const prodBranch = getInput("prod-branch").toLowerCase().trim();
-  const prodTag = getInput("prod-tag").toLowerCase().trim();
-
-  const shouldTagHead = tagHead === 'true' || tagHead === false;
-  if (shouldTagHead) {
-    const branchName = getBranchName();
-    if (branchName === prodBranch) {
-      headTag.push(`-t ${packagePath}:${prodTag}`)
-    } else if (branchName === stageBranch) {
-      headTag.push(`-t ${packagePath}:${stageTag}`)
-    } else {
-      headTag.push(`-t ${packagePath}:pr-${branchName}`);
-    }
-  }
-  return headTag;
-}
+const { setFailed, setOutput, debug } = require("@actions/core");
+const { getPushCommand, getBuildCommand, getLoginCommand, packagePath } = require('./docker');
 
 async function run() {
-  BUILD_ARGS.forEach(ENV => (setSecret(ENV)));
-  const token = getInput("repo-token");
-  const dockerfileLocation = getInput("dockerfile-location");
-  const imageName = getInput("image-name").toLowerCase();
-  const tag = getInput("tag").toLowerCase().trim().slice(0, 8);
-
-  const username = process.env.GITHUB_ACTOR;
-  const githubRepo = process.env.GITHUB_REPOSITORY.toLowerCase();
-
-  const packagePath = `docker.pkg.github.com/${githubRepo}/${imageName}`;
-  const dockerBuildArr = ['build'];
-
-  const parsedBuildArgs = parseBuildArgs();
-  dockerBuildArr.push(...parsedBuildArgs);
-
-  const parsedHeadTags = parseHeadTag(packagePath);
-  dockerBuildArr.push(...parsedHeadTags);
-
-  dockerBuildArr.push(`-t ${packagePath}:${tag}`);
-  dockerBuildArr.push(dockerfileLocation);
+  const login = getLoginCommand();
+  const build = getBuildCommand();
+  const push = getPushCommand()
 
   try {
-    await exec(
-      `docker login docker.pkg.github.com -u ${username} -p ${token}`
-    );
+    await exec(login);
   } catch (err) {
     setFailed(`action failed with error: ${err}`);
   }
   try {
-    debug(dockerBuildArr)
-    await exec(`docker ${dockerBuildArr.join(' ')}`);
+    debug(build)
+    await exec(build);
   } catch (err) {
     setFailed(`action failed with error: ${err}`);
   }
   try {
-    await exec(`docker push ${packagePath}`);
+    await exec(push);
   } catch (err) {
     setFailed(`Review the logs above, most likely you are using a package name associated with a different repository.  Rename your Image to fix. https://help.github.com/en/github/managing-packages-with-github-packages/about-github-packages#managing-packages for more information`);
   }
